Refatoring code to seperate modules according to functionalities o class
