--- src/main/java/com/sample/vivek/kafka/learning/exception/PropertyReadException.java
@@ -1,19 +1,0 @@
-package com.sample.vivek.kafka.learning.exception;
-
-/**
- * @author : Vivek Kumar Gupta
- * @since : 20/07/20
- */
-public class PropertyReadException extends Exception {
-
-    public PropertyReadException(String msg){
-        super(msg);
-    }
-
-    public PropertyReadException(String msg, Throwable exp){
-        super(msg,exp);
-    }
-
-
-
-}
